Completed UI portion of App, also added folder for forms , as well as added placeholder data to test binding of models to UI

+++ Models/Inhouse.cs
@@ -3,7 +3,7 @@
     public class Inhouse : Part
     {
         public int MachineID { get; set; }
-        public Inhouse(int partID, string name, double price, int inStock, int min, int max, int machineID)
+        public Inhouse(int partID, string name, decimal price, int inStock, int min, int max, int machineID)
             : base(partID, name, price, inStock, min, max)
         {
             MachineID = machineID;
