+++ Models/Outsourced.cs
@@ -3,7 +3,7 @@
     public class Outsourced : Part
     {
         public string CompanyName { get; set; }
-        public Outsourced(int partID, string name, double price, int inStock, int min, int max, string companyName)
+        public Outsourced(int partID, string name, decimal price, int inStock, int min, int max, string companyName)
             : base(partID, name, price, inStock, min, max)
         {
             CompanyName = companyName;
