+++ Models/Part.cs
@@ -5,12 +5,12 @@ namespace InventoryMgmt.Models
     {
         public int PartID { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int InStock { get; set; }
         public int Min { get; set; }
         public int Max { get; set; }
 
-        protected Part(int partID, string name, double price, int inStock, int min, int max)
+        protected Part(int partID, string name, decimal price, int inStock, int min, int max)
         {
             PartID = partID;
             Name = name;
